gemini-streamlit-cloudrun/retrieval_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
pydantic==2.6.4
pgvector==0.2.4
python-dotenv==1.0.1
